refactorign login
